bolsa-web-scraping/requirements.txt
Flask==2.0.1
datetime==4.3
requests==2.26.0
pytz==2021.1
google-cloud-storage==1.42.0